add angular2 generator proto, with js not ts for the moment
