+++ generators/angular2/templates/src/app/index.scss
@@ -1,0 +1,3 @@
+body {
+  background-color: lightgrey;
+}
