+++ generators/angular2/templates/src/index.html
@@ -1,0 +1,19 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>FountainJS</title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width">
+
+    <link rel="stylesheet" href="index.css"></link>
+  </head>
+
+  <body>
+    <hello-app>Loading...</hello-app>
+  </body>
+
+  <script src="/node_modules/reflect-metadata/Reflect.js"></script>
+  <script src="/node_modules/zone.js/dist/zone.js"></script>
+  <script src="index.js"></script>
+</html>
